Split functionality into lib and main crates
This should also make it easier to test, and easier to benchmark properly using
criterion.
