+++ rust/Cargo.toml
@@ -10,6 +10,7 @@ edition = "2018"
 rayon = "1.1"
 structopt = { version = "0.2.18", features = ["paw"] }
 paw = "1.0.0"
+lazy_static = "1.4.0"
 
 [profile.dev]
 opt-level = 3
